♻️ refactor: index.js 변경
